Adapt to MySQL support splitting out the timezone offset
Since the sender's timezone is in a separate column, we can now display
the message date in Django's timezone and in the sender's timezone.

Future improvement : ask the user for his/her preferred timezone and store
it in his/her profile. Then use that to display the dates, instead of
Django's server-wide setting.